scheduler(ticdc): add splitter interface (#8505)
ref pingcap/tiflow#7720
